Add react-router and new routes
