+++ package.json
@@ -11,10 +11,12 @@
   },
   "homepage": "https://github.com/harnyk/the-drift",
   "dependencies": {
+    "@types/react-router-dom": "^5.3.3",
     "clsx": "^2.1.1",
     "focus-trap-react": "^11.0.4",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
+    "react-router-dom": "^7.6.2",
     "vercel": "^42.3.0"
   },
   "devDependencies": {
