+++ src/ui/AboutPage.tsx
@@ -1,0 +1,9 @@
+import React from 'react';
+
+const AboutPage: React.FC = () => (
+    <div className="flex items-center justify-center h-screen">
+        <h1 className="text-4xl">The Drift</h1>
+    </div>
+);
+
+export default AboutPage;
