+++ src/ui/App.tsx
@@ -1,44 +1,15 @@
-import React, { useEffect, useState } from 'react';
-import LayeredLayout from './LayeredLayout';
-import GameCanvas from './GameCanvas';
-import PauseMenu from './PauseMenu';
-import { DialogProvider, useDialogManager } from './DialogManager';
-
-const InnerApp: React.FC = () => {
-    const [paused, setPaused] = useState(false);
-    const { hasDialog, showDialog, closeDialog } = useDialogManager();
-
-    useEffect(() => {
-        const onKeyDown = (e: KeyboardEvent) => {
-            if (e.code === 'Escape' && !hasDialog) {
-                setPaused(true);
-                showDialog(
-                    (id, isTop) => (
-                        <PauseMenu
-                            dialogId={id}
-                            isTop={isTop}
-                            onExit={() => closeDialog(id)}
-                        />
-                    ),
-                    () => setPaused(false)
-                );
-            }
-        };
-        window.addEventListener('keydown', onKeyDown);
-        return () => {
-            window.removeEventListener('keydown', onKeyDown);
-        };
-    }, [hasDialog, showDialog, closeDialog]);
-
-    return (
-        <LayeredLayout layers={[<GameCanvas key="game" paused={paused} />]} />
-    );
-};
+import React from 'react';
+import { BrowserRouter, Routes, Route } from 'react-router-dom';
+import GameApp from './GameApp';
+import AboutPage from './AboutPage';
 
 const App: React.FC = () => (
-    <DialogProvider>
-        <InnerApp />
-    </DialogProvider>
+    <BrowserRouter>
+        <Routes>
+            <Route path="/" element={<GameApp />} />
+            <Route path="/about" element={<AboutPage />} />
+        </Routes>
+    </BrowserRouter>
 );
 
 export default App;
